backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
openai==1.3.5
groq==0.4.1
python-docx==1.1.0
pydantic==2.5.0

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
pytest-mock==3.12.0
requests==2.31.0

